[Feature Set 9] Handle ICMP TDN change in a work queue.
Since changing tp->curr_tdn_id would require synchronization, the TDN
change handler needs to lock the socket when making this update. But
ICMP packets are handled in the softirq interrupt context where blocking
of any form is illegal. Hence it cannot wait to lock the socket,
otherwise it causes a deadlock. As an alternative, this change creates a
work queue, which allows blocking, to handle the TDN change.