[bug fix] Revert subflow total_retrans to global total_retrans.
Previously, TDTCP changed to use a per-TDN/subflow total_retrans counter
as part of the congestion related variables. But it turns out that
total_retrans is only used as statistics rather than in meaningful
congestion control. And iperf does not handle per-TDN total_retrans well
because it reports retransmissions in a time interval based on diff of
two reads of the variable. We have seen that retransmission report goes
negative because total_retrans[0] read at t is 1 but total_retrans[1]
read at t + 1 is 0, hence delta is -1.

Reverting all TDTCP changes related to this variable should fix the bug.