[Bug fix] Track retransmitted SKB with `SKB.retx_tdn_id`.
One bug that causes retrans_out to overflow is because
tcp_retransmit_skb() marks a retransmitted SKB on retrans_out of
curr_tdn_id but sack/ack handling subtracts it from retrans_out of
SKB.data_tdn_id. This inconsistency reveals that retransmission can
happen on a TDN different than the original one. Hence, we need to
introduce a `retx_tdn_id` for this use case.