[Feature Set 11] Refactor code that accesses and updates current TDN ID.
This change is to prepare for an optimized implementation of TDN update
handling. Current implementation associates a `curr_tdn_id` field with
each TDTCP-enabled socket and allows each socket to have an independent
TDN ID. This however creates performance issues when there are many
TDTCP sockets to be updated simultaneously. Based on measurements, in a
*-lowlatency kernel, reacting to TDN change takes 500 usec per socket.
Thus, it won't be able for all TDTCP sockets to keep up with TDN updates
faster than 500 usec/socket * N sockets.

A new optimized implementation introduces a global TDN ID that will be
shared with all TDTCP sockets. Updating this TDN ID is O(1) complexity.