[Bug fix] Assign data_tdn_id for SYN SKB.
data_tdn_id was only set in tcp_write_xmit() for pure data SKBs. SYN
SKBs have an unintialized data_tdn_id which defaults to 0. When doing
accounting, pkts_out of the current TDN is updated since it is unclear
which TDN the SYN SKB belongs to. This is now fixed with a properly
initialized SYN data_tdn_id.